Feature Main Refactor (#135)
* Pull implementation details out of main and into their own classes.

* Fix mocks in some tests.
