+++ webbreaker/__init__.py
@@ -1,1 +1,2 @@
-__version__ = '2.1.4'
+__version__ = '2.1.5'
+
